Achieved all base functionality
There are now several NPCs to talk to, with a chain of item trades leading to a boat ticket. I made some minor tweaks to the NPCController script to allow toggling a collider as a door. I also integrated some preliminary character art and made some mock-up items with primitives.